+++ CHANGELOG.rst
@@ -23,6 +23,9 @@ Changelog
 * Support serialization of certificate revocation lists using the
   :meth:`~cryptography.x509.CertificateRevocationList.public_bytes` method of
   :class:`~cryptography.x509.CertificateRevocationList`.
+* Add support for parsing :class:`~cryptography.x509.CertificateRevocationList`
+  :meth:`~cryptography.x509.CertificateRevocationList.extensions` in the
+  OpenSSL backend.
 
 1.1.2 - 2015-12-10
 ~~~~~~~~~~~~~~~~~~